# web/static/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_alchemillahq_sylve_internal.APIResponse-any:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  ? github_com_alchemillahq_sylve_internal.APIResponse-github_com_alchemillahq_sylve_internal_interfaces_services_zfs_SimpleZFSDiskUsage
  : properties:
      data:
        $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.SimpleZFSDiskUsage'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  github_com_alchemillahq_sylve_internal.APIResponse-gzfs_ZPoolStatusPool:
    properties:
      data:
        $ref: '#/definitions/gzfs.ZPoolStatusPool'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal.APIResponse-internal_handlers_zfs_PoolStatPointResponse:
    properties:
      data:
        $ref: '#/definitions/internal_handlers_zfs.PoolStatPointResponse'
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
  github_com_alchemillahq_sylve_internal_db.IntervalOption:
    properties:
      label:
        type: string
      value:
        type: integer
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.CreateZPoolRequest:
    properties:
      createForce:
        type: boolean
      name:
        maxLength: 24
        minLength: 1
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
      raidType:
        $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.RaidType'
      spares:
        items:
          type: string
        type: array
      vdevs:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.Vdev'
        type: array
    required:
    - name
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.PoolStatPoint:
    properties:
      allocated:
        type: integer
      dedupRatio:
        type: number
      free:
        type: integer
      size:
        type: integer
      time:
        type: integer
    type: object
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.RaidType:
    enum:
    - stripe
    - mirror
    - raidz
    - raidz2
    - raidz3
    type: string
    x-enum-varnames:
    - RaidTypeStripe
    - RaidTypeMirror
    - RaidTypeRaidZ
    - RaidTypeRaidZ2
    - RaidTypeRaidZ3
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.ReplaceDevice:
    properties:
      new:
        maxLength: 24
        minLength: 1
        type: string
      old:
        maxLength: 24
        minLength: 1
        type: string
    required:
    - new
    - old
    type: object
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.SimpleZFSDiskUsage:
    properties:
      total:
        type: number
      usage:
        type: number
    type: object
...
  github_com_alchemillahq_sylve_internal_interfaces_services_zfs.Vdev:
    properties:
      devices:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
...
  gzfs.ZFSProperty:
    properties:
      source:
        $ref: '#/definitions/gzfs.ZFSPropertySource'
      value:
        type: string
    type: object
  gzfs.ZFSPropertySource:
    properties:
      data:
        type: string
      type:
        type: string
    type: object
  gzfs.ZPool:
    properties:
      allocated:
        type: integer
      dedup_ratio:
        type: number
      fragmentation:
        type: number
      free:
        type: integer
      l2cache:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolVDEV'
        type: object
      logs:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolVDEV'
        type: object
      name:
        type: string
      pool_guid:
        type: string
      properties:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZFSProperty'
        type: object
      size:
        type: integer
      spa_version:
        type: string
      spares:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolVDEV'
        type: object
      state:
        $ref: '#/definitions/gzfs.ZPoolState'
      txg:
        type: string
      type:
        type: string
      zpl_version:
        type: string
    type: object
  gzfs.ZPoolState:
    enum:
    - ONLINE
    - DEGRADED
    - FAULTED
    - OFFLINE
    - REMOVED
    - UNAVAIL
    - CORRUPT_DATA
    - UNKNOWN
    type: string
    x-enum-varnames:
    - ZPoolStateOnline
    - ZPoolStateDegraded
    - ZPoolStateFaulted
    - ZPoolStateOffline
    - ZPoolStateRemoved
    - ZPoolStateUnavailible
    - ZPoolStateCorruptData
    - ZPoolStateUnknown
  gzfs.ZPoolStatusPool:
    properties:
      action:
        type: string
      l2cache:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolStatusVDEV'
        type: object
      logs:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolStatusVDEV'
        type: object
      name:
        type: string
      pool_guid:
        type: string
      scan_stats:
        $ref: '#/definitions/gzfs.ZPoolStatusScanStats'
      spa_version:
        type: string
      spares:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolStatusVDEV'
        type: object
      state:
        type: string
      status:
        type: string
      txg:
        type: string
      vdevs:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolStatusVDEV'
        type: object
      zpl_version:
        type: string
    type: object
  gzfs.ZPoolStatusScanStats:
    properties:
      bytes_per_scan:
        type: string
      end_time:
        type: string
      errors:
        type: string
      examined:
        type: string
      function:
        type: string
      issued:
        type: string
      issued_bytes_per_scan:
        type: string
      pass_start:
        type: string
      processed:
        type: string
      scrub_pause:
        type: string
      scrub_spent_paused:
        type: string
      skipped:
        type: string
      start_time:
        type: string
      state:
        type: string
      to_examine:
        type: string
    type: object
  gzfs.ZPoolStatusVDEV:
    properties:
      alloc_space:
        type: string
      checksum_errors:
        type: string
      class:
        type: string
      def_space:
        type: string
      guid:
        type: string
      name:
        type: string
      path:
        type: string
      read_errors:
        type: string
      rep_dev_size:
        type: string
      state:
        type: string
      total_space:
        type: string
      vdev_type:
        type: string
      vdevs:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolStatusVDEV'
        type: object
      write_errors:
        type: string
    type: object
  gzfs.ZPoolVDEV:
    properties:
      allocated:
        type: integer
      class:
        type: string
      fragmentation:
        type: number
      free:
        type: integer
      guid:
        type: string
      name:
        type: string
      path:
        type: string
      phys_path:
        type: string
      properties:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZFSProperty'
        type: object
      size:
        type: integer
      state:
        type: string
      vdev_type:
        type: string
      vdevs:
        additionalProperties:
          $ref: '#/definitions/gzfs.ZPoolVDEV'
        type: object
    type: object
...
  internal_handlers_zfs.PoolEditRequest:
    properties:
      name:
        type: string
      properties:
        additionalProperties:
          type: string
        type: object
      spares:
        items:
          type: string
        type: array
    type: object
  internal_handlers_zfs.PoolStatPointResponse:
    properties:
      intervalMap:
        items:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_db.IntervalOption'
        type: array
      poolStatPoint:
        additionalProperties:
          items:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.PoolStatPoint'
          type: array
        type: object
    type: object
...
  internal_handlers_zfs.ZpoolListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/gzfs.ZPool'
        type: array
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
...
host: sylve.lan:8181
info:
  contact:
    email: hello@alchemilla.io
    name: Alchemilla Ventures Pvt. Ltd.
    url: https://alchemilla.io
  description: Sylve is a lightweight GUI for managing Bhyve, Jails, ZFS, networking,
    and more on FreeBSD.
  license:
    name: BSD-2-Clause
    url: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  termsOfService: https://github.com/AlchemillaHQ/Sylve/blob/master/LICENSE
  title: Sylve API
  version: 0.0.1
paths:
...
  /zfs/pool/stats/{interval}/{limit}:
    get:
      consumes:
      - application/json
      description: Get the historical stats of a ZFS pool
      parameters:
      - description: Interval in minutes
        in: path
        name: interval
        required: true
        type: integer
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-internal_handlers_zfs_PoolStatPointResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Pool Stats
      tags:
      - ZFS
  /zfs/pools:
    get:
      consumes:
      - application/json
      description: Get all ZFS pools
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/internal_handlers_zfs.ZpoolListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Pools
      tags:
      - ZFS
    patch:
      consumes:
      - application/json
      description: Edit a ZFS pool's properties
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handlers_zfs.PoolEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Edit Pool
      tags:
      - ZFS
    post:
      consumes:
      - application/json
      description: Create a new ZFS pool
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.CreateZPoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Create Pool
      tags:
      - ZFS
  /zfs/pools/{guid}:
    delete:
      consumes:
      - application/json
      description: Delete a ZFS pool
      parameters:
      - description: Pool GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Delete Pool
      tags:
      - ZFS
  /zfs/pools/{guid}/replace-device:
    post:
      consumes:
      - application/json
      description: Replace a device in a ZFS pool
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alchemillahq_sylve_internal_interfaces_services_zfs.ReplaceDevice'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Replace Device
      tags:
      - ZFS
  /zfs/pools/{guid}/scrub:
    post:
      consumes:
      - application/json
      description: Start a scrub on a ZFS pool
      parameters:
      - description: Pool GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Scrub Pool
      tags:
      - ZFS
  /zfs/pools/{guid}/status:
    get:
      consumes:
      - application/json
      description: Get the status of a ZFS pool
      parameters:
      - description: Pool GUID
        in: path
        name: guid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-gzfs_ZPoolStatusPool'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Pool Status
      tags:
      - ZFS
  /zfs/pools/disk-usage:
    get:
      consumes:
      - application/json
      description: Get the overall disk usage percentage across all ZFS pools
      produces:
      - application/json
      responses:
        "200":
          description: Disk usage percentage
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-github_com_alchemillahq_sylve_internal_interfaces_services_zfs_SimpleZFSDiskUsage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alchemillahq_sylve_internal.APIResponse-any'
      security:
      - BearerAuth: []
      summary: Get Disk Usage
      tags:
      - ZFS
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"